removed redundancy of some methods in staging environment
